add lots of QOL improvements; game balance; bugfixes
- fix centipede getting stuck on environment
- fix player state reset on level warp
- fix bullets sticking to objects
- tweak credits speed
- add door sprites
- fix hazard and boundary tiles
- add health to boss
- tweak boss cinematic
- increase player movement speed slightly
- add randomness to static cowboy shot
