+++ Assets/Animations/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 208fe53b255c0bc49b149f852d08e522
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
