+++ Assets/Animations/Environment/Door.anim
@@ -1,0 +1,107 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Door
+  serializedVersion: 6
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - curve:
+    - time: 0
+      value: {fileID: -4294340948301892083, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.083333336
+      value: {fileID: 3741775267428246300, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.16666667
+      value: {fileID: 7920950194370322701, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.25
+      value: {fileID: 5096838038997782895, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.33333334
+      value: {fileID: -1589554221223979101, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.41666666
+      value: {fileID: 4181975277439144173, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.5
+      value: {fileID: 7650364836354595358, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.5833333
+      value: {fileID: -5785852764959157235, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.6666667
+      value: {fileID: -1497538912974577852, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.75
+      value: {fileID: -2249309058153822101, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.8333333
+      value: {fileID: -551142303996023719, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 0.9166667
+      value: {fileID: -4188855112023833423, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 1
+      value: {fileID: -8183702973515052970, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - time: 1.0833334
+      value: {fileID: 7705600931651263846, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+  m_SampleRate: 12
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+    pptrCurveMapping:
+    - {fileID: -4294340948301892083, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 3741775267428246300, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 7920950194370322701, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 5096838038997782895, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -1589554221223979101, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 4181975277439144173, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 7650364836354595358, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -5785852764959157235, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -1497538912974577852, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -2249309058153822101, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -551142303996023719, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -4188855112023833423, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: -8183702973515052970, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+    - {fileID: 7705600931651263846, guid: e560b20991d25aa4faa95ea14287bab2, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.1666667
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 1
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events: []
