+++ Assets/Animations/Environment/Door.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07d5732903fbdbe44a3105b777aca631
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
