+++ Assets/Animations/Environment/door.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8e69bc12875091459509d65799c1856
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
