--- Assets/Boss.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b052cf5e1ef53b549a9ef36d1b4333f5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
