--- Assets/CowboyAnimationAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f4cde8d80b19f3e459c77cdf3e06fbc3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
