+++ Assets/Lib/TheKiwiCoder/BehaviourTree/Scripts/Runtime/NodePort.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 
 using System.Collections.Generic;
 using UnityEngine;
@@ -68,3 +69,5 @@ namespace TheKiwiCoder {
         }
     }
 }
+
+#endif
