+++ Assets/Prefabs/Bullets/PlayerBullet.prefab
@@ -53,7 +53,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 8673064300862581254, guid: 42aec47ca2ab376439cca64f2ce5e227, type: 3}
       propertyPath: m_Name
-      value: BulletProto
+      value: PlayerBullet
+      objectReference: {fileID: 0}
+    - target: {fileID: 8673064300862581261, guid: 42aec47ca2ab376439cca64f2ce5e227, type: 3}
+      propertyPath: debug
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 8673064300862581261, guid: 42aec47ca2ab376439cca64f2ce5e227, type: 3}
+      propertyPath: ignoreTag
+      value: Player
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 42aec47ca2ab376439cca64f2ce5e227, type: 3}
