+++ Assets/Prefabs/Enemies/StaticCowboy.prefab
@@ -11,6 +11,14 @@ PrefabInstance:
       propertyPath: m_Controller
       value: 
       objectReference: {fileID: 9100000, guid: e9ac51ffe8e485c4abdb6675318f116e, type: 2}
+    - target: {fileID: 6565066346141863061, guid: b901b3530da270741842c1614b8e6557, type: 3}
+      propertyPath: accuracyAngle
+      value: 10
+      objectReference: {fileID: 0}
+    - target: {fileID: 6565066346141863061, guid: b901b3530da270741842c1614b8e6557, type: 3}
+      propertyPath: hasPerfectAccuracy
+      value: 0
+      objectReference: {fileID: 0}
     - target: {fileID: 6565066347014586354, guid: b901b3530da270741842c1614b8e6557, type: 3}
       propertyPath: m_Radius
       value: 0.64
