+++ Assets/Prefabs/PlayerAndSystems/Player.prefab
@@ -301,7 +301,7 @@ CircleCollider2D:
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
   serializedVersion: 2
-  m_Radius: 2.5
+  m_Radius: 3
 --- !u!114 &3844958333541571679
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -318,7 +318,7 @@ MonoBehaviour:
   damageAmount: 20
   damageVariance: 4
   makeFramerateIndependent: 0
-  impactForce: 5
+  impactForce: 10
   ignoreParentGUID: 1
   ignoreLayers:
     serializedVersion: 2
@@ -618,7 +618,7 @@ Transform:
   m_GameObject: {fileID: 3844958334892354028}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 5, y: 5, z: 1}
+  m_LocalScale: {x: 6, y: 6, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 3844958333541571677}
@@ -982,7 +982,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7bf0e0764d90e9b41b247c8d48deb33e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  maxSpeed: 5
+  maxSpeed: 6
   throttleUpTime: 0.5
   throttleDownTime: 0.05
   speedDelta: 0.5
@@ -1002,7 +1002,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   startingHP: 100
   isInvulnerable: 0
-  timeInvincibleAfterHit: 0.1
+  timeInvincibleAfterHit: 0.4
 --- !u!114 &3844958334926481135
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1044,7 +1044,7 @@ MonoBehaviour:
   timeRecharge: 3
   numberOfShots: 5
   timeIsShootingThreshold: 0.1
-  timeIsMeleeingThreshold: 0.6
+  timeIsMeleeingThreshold: 0.3
   eventChannel: {fileID: 11400000, guid: 7981a9b65533ac444b7650afd7b5691d, type: 2}
 --- !u!114 &3844958334926485777
 MonoBehaviour:
