+++ Assets/Scripts/Actor/Health.cs
@@ -55,6 +55,7 @@ namespace Actor {
             if (!isAlive) return false;
             if (damage <= 0) return false;
             if (isInvulnerable && damage != Constants.INSTAKILL) return false;
+            if (timeInvincible.active && damage != Constants.INSTAKILL) return false;
             if (hp <= 0) return false;
 
             hp -= damage;
