+++ Assets/Scripts/BTree/Actions/ChasePlayerAction.cs
@@ -7,10 +7,12 @@ public class ChasePlayerAction : ActionNode {
         if (Player.PlayerUtils.CheckIsAlive(context.player)) {
             context.movement.SetTarget(context.player.transform);
         }
+        if (context.wander != null) context.wander.enabled = false;
     }
 
     protected override void OnStop() {
         context.movement.Halt();
+        if (context.wander != null) context.wander.enabled = true;
     }
 
     protected override State OnUpdate() {
