+++ Assets/Scripts/InvertParentRotation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0414290d38a86946a0a51e54164bb50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
