+++ Assets/Sprites/door.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7571aa15bcc0c5248b60db7d5c43a373
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
