+++ Assets/Tiles/Lava.asset
@@ -26,4 +26,4 @@ MonoBehaviour:
   m_MaxSpeed: 1
   m_AnimationStartTime: 0
   m_AnimationStartFrame: 0
-  m_TileColliderType: 0
+  m_TileColliderType: 1
