+++ Assets/Tiles/doorTile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7eee8c3f87e6f34095b46756182b51c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
