+++ Packages/manifest.json
@@ -1,4 +1,13 @@
 {
+  "scopedRegistries": [
+      {
+          "name": "package.openupm.com",
+          "url": "https://package.openupm.com",
+          "scopes": [
+              "io.tinu.asepriteimporter"
+          ]
+      }
+  ],
   "dependencies": {
     "com.unity.cinemachine": "2.8.6",
     "com.unity.collab-proxy": "1.17.1",
@@ -42,6 +51,7 @@
     "com.unity.modules.video": "1.0.0",
     "com.unity.modules.vr": "1.0.0",
     "com.unity.modules.wind": "1.0.0",
-    "com.unity.modules.xr": "1.0.0"
-  }
+    "com.unity.modules.xr": "1.0.0",
+    "io.tinu.asepriteimporter": "1.0.4"
+  }   
 }
