+++ Packages/packages-lock.json
@@ -241,6 +241,15 @@
       },
       "url": "https://packages.unity.com"
     },
+    "io.tinu.asepriteimporter": {
+      "version": "1.0.4",
+      "depth": 0,
+      "source": "registry",
+      "dependencies": {
+        "com.unity.2d.sprite": "1.0.0"
+      },
+      "url": "https://package.openupm.com"
+    },
     "com.unity.modules.ai": {
       "version": "1.0.0",
       "depth": 0,
